feat: remove unecessary files
