--- frontend/src/assets/index.js
@@ -1,9 +1,0 @@
-import download from "./download.png";
-import aiqem_logo from "./aiqem_logo.svg";
-import logo from "./aiqem_logo.svg";
-import preview from "./preview.png";
-import hero from "./hero.png";
-import africa1 from "./africa1.jpg";
-import images from "./images.jpeg"
-import telegram from "./telegram.jpeg";
-export { download, aiqem_logo, logo,  preview, hero, africa1 , images, telegram};
