--- frontend/src/components/index.js
@@ -1,5 +1,0 @@
-import Card from "./Card";
-import FormFields from "./FormFields";
-import Loader from "./Loader";
-
-export { Card, FormFields, Loader };
