--- frontend/src/pages/index.js
@@ -1,4 +1,0 @@
-import Home from "./Home";
-import CreatePost from "./CreatePost";
-
-export { Home, CreatePost };
